+++ src/core/components/Linkify/styles.js
@@ -1,9 +1,10 @@
 import styled from 'styled-components';
 
 export const Link = styled.a`
-  &,
-  &:visited {
-    color: ${({ theme }) => theme.palette.primary.main};
-    text-decoration: none;
+  color: ${({ theme }) => theme.palette.primary.main};
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
   }
 `;