+++ src/social/components/UserInfo/index.js
@@ -9,6 +9,7 @@ import { useNavigation } from '~/social/providers/NavigationProvider';
 import UIUserInfo from './UIUserInfo';
 import useFollow from '~/core/hooks/useFollow';
 import useFollowCount from '~/core/hooks/useFollowCount';
+import useImage from '~/core/hooks/useImage';
 import { useAsyncCallback } from '~/core/hooks/useAsyncCallback';
 import { notification } from '~/core/components/Notification';
 import { UserFeedTabs } from '~/social/pages/UserFeed/constants';
@@ -30,7 +31,9 @@ const UserInfo = ({
   const { followerCount, followingCount } = useFollowCount(userId);
 
   const { displayName, description } = user;
-  const { fileUrl } = file;
+
+  const { fileId } = file;
+  const fileUrl = useImage({ fileId });
 
   const [onFollowDecline] = useAsyncCallback(async () => {
     await followDecline();
@@ -42,7 +45,7 @@ const UserInfo = ({
     <UIUserInfo
       userId={userId}
       currentUserId={currentUserId}
-      fileUrl={fileUrl}
+      fileUrl={user.avatarCustomUrl || fileUrl}
       displayName={displayName || formatMessage({ id: 'anonymous' })}
       description={description}
       setActiveTab={setActiveTab}