+++ src/social/components/post/Creator/components/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import UIAvatar from '~/core/components/Avatar';
+
+export const Avatar = styled(UIAvatar)`
+  margin-right: 8px;
+`;